This test deserves a separate file
